feat(cart): add functionality to remove products from cart and update… (#13)
* feat(cart): add functionality to remove products from cart and update cart totals

* - Add validation guards for negative cart totals
- Migrate Double to BigDecimal for price calculations
- Implement idempotent DELETE (return success if product already removed)
- Fix typo: succesfully → successfully
- Add detailed error logging for corrupted cart state"

+++ MyData-service/src/main/java/Ecommerce/Reactive/MyData_service/DTO/cartProducts/ResponseProductToCartDto.java
@@ -3,6 +3,8 @@ package Ecommerce.Reactive.MyData_service.DTO.cartProducts;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ResponseProductToCartDto {
@@ -10,12 +12,12 @@ public class ResponseProductToCartDto {
     // details of the product added
     private Long id;
     private String name;
-    private Double price;
+    private BigDecimal price;
 
     // actual cart details
     private Long cartId;
     private Integer cartTotalItems;
-    private Double cartTotalPrice;
+    private BigDecimal cartTotalPrice;
 
     private String message;
 }
